requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
requests
psycopg2-binary
supabase

